xbeehub/requirements.txt
XBee==2.1.0
argparse==1.2.1
epc==0.0.5
jedi==0.9.0
paho-mqtt==1.1
sexpdata==0.0.3
wsgiref==0.1.2
